fixup! Closes #11 - Adds outages api implementation

+++ apps/microservices/ms-metrics/src/main/java/br/com/example/ms/metrics/entrypoint/rest/MetaView.java
@@ -7,5 +7,5 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class MetaView {
 	private final int totalPages;
-	private final int totalRecords;
+	private final long totalRecords;
 }
